Improve google_project#to_s output
The google_project resource documentation indicates that resources are
instantiated with :project, but the google_project was storing the value
of :name for use with `#to_s`. This inconsistency leads to InSpec output
like the following:

     ✔  Project  should exist
     ✔  Project  name should eq "dev-host"
     ✔  Project  parent.type should eq "folder"
     ✔  Project  parent.id should eq "743996818495"

This commit adds :project as a fallback value for `@display_name,`
ensuring that google_projects will format correctly in the InSpec
report.

Signed-off-by: Adrien Thebo <adrien@lagrange-automation.io>